[AdaGFX] Correct DisplayTimeout setting not allow negative, and > 255 seconds

+++ src/src/PluginStructs/P095_data_struct.cpp
@@ -22,7 +22,7 @@ const __FlashStringHelper* P095_CommandTrigger_toString(P095_CommandTrigger cmd)
 P095_data_struct::P095_data_struct(uint8_t             rotation,
                                    uint8_t             fontscaling,
                                    AdaGFXTextPrintMode textmode,
-                                   uint8_t             displayTimer,
+                                   uint32_t            displayTimer,
                                    String              commandTrigger,
                                    uint16_t            fgcolor,
                                    uint16_t            bgcolor,
@@ -317,10 +317,10 @@ bool P095_data_struct::plugin_write(struct EventStruct *event, const String& str
 /****************************************************************************
  * displayOnOff: Turn display on or off
  ***************************************************************************/
-void P095_data_struct::displayOnOff(bool    state,
-                                    int8_t  backlightPin,
-                                    uint8_t backlightPercentage,
-                                    uint8_t displayTimeout) {
+void P095_data_struct::displayOnOff(bool     state,
+                                    int8_t   backlightPin,
+                                    uint8_t  backlightPercentage,
+                                    uint32_t displayTimeout) {
   if (backlightPin != -1) {
     # if defined(ESP8266)
     analogWrite(backlightPin, state ? ((1024 / 100) * backlightPercentage) : 0);
