+++ src/src/PluginStructs/P095_data_struct.h
@@ -73,7 +73,7 @@ public:
   P095_data_struct(uint8_t             rotation,
                    uint8_t             fontscaling,
                    AdaGFXTextPrintMode textmode,
-                   uint8_t             displayTimer,
+                   uint32_t            displayTimer,
                    String              commandTrigger,
                    uint16_t            fgcolor      = ADAGFX_WHITE,
                    uint16_t            bgcolor      = ADAGFX_BLACK,
@@ -96,10 +96,10 @@ public:
 
 private:
 
-  void displayOnOff(bool    state,
-                    int8_t  backlightPin,
-                    uint8_t backlightPercentage,
-                    uint8_t displayTimeout);
+  void displayOnOff(bool     state,
+                    int8_t   backlightPin,
+                    uint8_t  backlightPercentage,
+                    uint32_t displayTimeout);
   void updateFontMetrics();
 
   Adafruit_ILI9341 *tft = nullptr;
@@ -117,7 +117,7 @@ private:
   uint8_t             _rotation;
   uint8_t             _fontscaling;
   AdaGFXTextPrintMode _textmode;
-  uint8_t             _displayTimer;
+  uint32_t            _displayTimer;
   String              _commandTrigger;
   uint16_t            _fgcolor;
   uint16_t            _bgcolor;
