+++ src/src/PluginStructs/P116_data_struct.h
@@ -98,7 +98,7 @@ public:
                    uint8_t             rotation,
                    uint8_t             fontscaling,
                    AdaGFXTextPrintMode textmode,
-                   uint8_t             displayTimer,
+                   uint32_t            displayTimer,
                    String              commandTrigger,
                    uint16_t            fgcolor      = ADAGFX_WHITE,
                    uint16_t            bgcolor      = ADAGFX_BLACK,
@@ -121,10 +121,10 @@ public:
 
 private:
 
-  void displayOnOff(bool    state,
-                    int8_t  backlightPin,
-                    uint8_t backlightPercentage,
-                    uint8_t displayTimeout);
+  void displayOnOff(bool     state,
+                    int8_t   backlightPin,
+                    uint8_t  backlightPercentage,
+                    uint32_t displayTimeout);
   void updateFontMetrics();
 
   Adafruit_ST77xx      *st77xx    = nullptr;
@@ -145,7 +145,7 @@ private:
   uint8_t             _rotation;
   uint8_t             _fontscaling;
   AdaGFXTextPrintMode _textmode;
-  uint8_t             _displayTimer;
+  uint32_t            _displayTimer;
   String              _commandTrigger;
   uint16_t            _fgcolor;
   uint16_t            _bgcolor;
